requirements.txt
transformers==4.53.1
datasets>=2.14.0
seqeval>=1.2.2
scikit-learn>=1.3.0
PyYAML>=6.0
numpy>=1.23
tokenizers>=0.13.3
tqdm>=4.66.0
torch>=2.0.0